Move the 'Add shortlisting methods' page
